+++ packages/wonder-blocks-layout/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/wonder-blocks-layout
 
+## 3.1.57
+
+### Patch Changes
+
+- Updated dependencies [9d8e5e3]
+    - @khanacademy/wonder-blocks-tokens@17.0.0
+
 ## 3.1.56
 
 ### Patch Changes