+++ packages/wonder-blocks-search-field/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @khanacademy/wonder-blocks-search-field
 
+## 5.1.77
+
+### Patch Changes
+
+- Updated dependencies [f1cd1ce]
+- Updated dependencies [9d8e5e3]
+- Updated dependencies [9d8e5e3]
+    - @khanacademy/wonder-blocks-icon@5.3.20
+    - @khanacademy/wonder-blocks-tokens@17.0.0
+    - @khanacademy/wonder-blocks-typography@5.0.0
+    - @khanacademy/wonder-blocks-form@7.6.7
+    - @khanacademy/wonder-blocks-icon-button@11.4.2
+
 ## 5.1.76
 
 ### Patch Changes