+++ packages/wonder-blocks-tooltip/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @khanacademy/wonder-blocks-tooltip
 
+## 4.1.83
+
+### Patch Changes
+
+- Updated dependencies [9d8e5e3]
+- Updated dependencies [9d8e5e3]
+    - @khanacademy/wonder-blocks-tokens@17.0.0
+    - @khanacademy/wonder-blocks-typography@5.0.0
+    - @khanacademy/wonder-blocks-layout@3.1.57
+    - @khanacademy/wonder-blocks-modal@8.7.8
+
 ## 4.1.82
 
 ### Patch Changes